move and promote Ray2d
